+++ src/Lint/RegexLintRequest.php
@@ -32,7 +32,7 @@ final readonly class RegexLintRequest
         public int $minSavings,
         private array $disabledSources = [],
         public bool $checkValidation = true,
-        public bool $checkRedos = true,
+        public bool $checkRedos = false,
         public bool $checkOptimizations = true,
         public int $analysisWorkers = 1,
         public array $optimizations = [],